Fix some warnings

+++ z.c
@@ -4,7 +4,7 @@
 #include "pmf.h"
 #include "z.h"
 
-struct pmf *z_generate(const struct pmf *t, int ts)
+struct pmf *z_generate(const struct pmf *t, unsigned int ts)
 {
   double sum;
   unsigned int j, th;
@@ -14,7 +14,7 @@ struct pmf *z_generate(const struct pmf *t, int ts)
 
   th = ts;
   sum = 0;
-  for (j = pmf_min(t); j < pmf_max(t); j++) {
+  for (j = (unsigned int)pmf_min(t); j < (unsigned int)pmf_max(t); j++) {
     if (j < ts) {
       fprintf(stderr, "Error!!! Interarrival times MUST be greater than %d!!!\n",
 		      ts);
